Add management interface using async networking
Traditional opensips modules use Linux shared memory to communicate
between different processes. Shared memory is not something I've dealt
with before, and it's not the most common thing to use in Rust.

Instead, I took the opportunity to demonstrate using spinning up a new
thread and running asynchronous networking code via Tokio. The parent
process creates a UNIX domain socket and awaits connections. Each
worker connects back to the parent.

When the management interface is triggered, that worker sends a packet
to the parent process, which then broadcasts the packet back out to
all of the children. When the children receive the packet, they
increment a counter.

To see this, you can perform an `OPTIONS` request, trigger the
management interface, then perform the `OPTIONS` request again. Note
that the counter value returned in the `OK` response changes:

```
$ sipsak -vv -s sip:127.0.0.1

   SIP/2.0 200 OK (hello world / 42 / 0)

$ docker exec -it opensips opensips-cli -x mi rust_experiment_control

"OK"

$ sipsak -vv -s sip:127.0.0.1

   SIP/2.0 200 OK (hello world / 42 / 1)
```